requirements.txt
APScheduler==3.6.3
future==0.18.2
futures==3.1.1
iso8601==0.1.12
modbus-tk==1.1.1
paho-mqtt==1.5.0
pkg-resources==0.0.0
pyserial==3.4
pytz==2020.1
PyYAML==5.3.1
serial==0.0.97
six==1.15.0
tzlocal==2.1
